requirements.txt
# lightning
pytorch-lightning
numpy
scipy
scikit-learn
einops
matplotlib
wandb
omegaconf
librosa
halo
pytest
